hello-world/07 multisegment display in rust
* 1Hz timer handles display data updates
* 1kHz timer handles LED refresh
* uses timers and manually compares
* added simpl utility: dump-avr-timers.py

    Useful for calculating which timers (and compare match register values)
    can be used for a given frequency ISR.

    Example:

        $ ./dump-avr-timers.py 1 100 1000
        # ------------------------------------
        # interrupt frequency: f
        # prescaler: p
        # compare match register value: cmr
        # timers: t
        # ------------------------------------
        "f: 1, p: 1, cmr: 15999999.0, t: None"
        "f: 1, p: 8, cmr: 1999999.0, t: None"
        "f: 1, p: 64, cmr: 249999.0, t: None"
        "f: 1, p: 256, cmr: 62499.0, t: [1]"
        "f: 1, p: 1024, cmr: 15624.0, t: [1]"
        "f: 100, p: 1, cmr: 159999.0, t: None"
        "f: 100, p: 8, cmr: 19999.0, t: [1]"
        "f: 100, p: 64, cmr: 2499.0, t: [1]"
        "f: 100, p: 256, cmr: 624.0, t: [1]"
        "f: 100, p: 1024, cmr: 155.25, t: [0, 2]"
        "f: 1000, p: 1, cmr: 15999.0, t: [1]"
        "f: 1000, p: 8, cmr: 1999.0, t: [1]"
        "f: 1000, p: 64, cmr: 249.0, t: [0, 2]"
        "f: 1000, p: 256, cmr: 61.5, t: [0, 2]"
        "f: 1000, p: 1024, cmr: 14.625, t: [0, 2]"

    This tells me that for a 1Hz frequency interrupt, I need to use timer1
    prescaled to 256 or 1024, and what the CMR values are for each. Similar
    values for the 100Hz and the 1kHz are shown. The dump command takes a
    list of frequencies on the command line.

+++ posts/hello-world/07-multisegment-led-rust-arduino/code/multisegment-led/avr/mseg-lib/src/bits.rs
@@ -1,8 +1,5 @@
-///! TODO: 1. introduce generics so that the input array
-///! TODO: can be any size. currently hardcoded to usize.
-///!
-///! TODO: 2. should I put in bounds checking? Right now the index could
-///! TODO: rollover if we leftshift more than the 8 bits in usize...
+///! TODO: should I put in bounds checking? Right now the index could
+///! TODO: rollover if we leftshift more than the number bits in T.
 ///!
 ///! Notes/Links:
 ///! * https://immunant.com/blog/2020/01/bitfields/
@@ -28,8 +25,8 @@
 /// let b4 = set(b2, 7, false);
 /// assert_eq!(0b00000001, b4);
 /// ```
-pub fn set(input: usize, index: usize, value: bool) -> usize {
-    let mask: usize = 1 << index;
+pub fn set(input: u8, index: usize, value: bool) -> u8 {
+    let mask: u8 = 1 << index;
     return if value {
         input | mask
     } else {
@@ -50,8 +47,8 @@ pub fn set(input: usize, index: usize, value: bool) -> usize {
 /// let b2 = get(0b00000001, 1);
 /// assert_eq!(b2, false);
 /// ```
-pub fn get(input: usize, index: usize) -> bool {
-    let mask: usize = 1 << index;
+pub fn get(input: u8, index: usize) -> bool {
+    let mask: u8 = 1 << index;
     let mut result = input & mask;
     result >>= index;
     return if result == 1 { true } else { false };
@@ -62,7 +59,7 @@ mod tests {
     #[test]
     /// set each bit on, one at a time, and verify that all the other bits are off.
     fn test_bits_set() {
-        let b: usize = 0b00000000;
+        let b: u8 = 0b00000000;
         use super::*;
 
         for i in 0..8 {
